Refactor which packages a bunch of controls are in to be a little bit more deliberate
