+++ src/xbot/common/controls/Nav6Gyro.java
@@ -1,9 +1,9 @@
 package xbot.common.controls;
 
 import edu.wpi.first.wpilibj.SerialPort;
+import xbot.common.controls.nav6.IMU;
 import xbot.common.math.ContiguousDouble;
 import xbot.common.wpi_extensions.mechanism_wrappers.XGyro;
-import xbot.nav6.IMU;
 
 /**
  * Internal class to handle gyro readings.
