+++ src/xbot/common/injection/wpi_factories/RealWPIFactory.java
@@ -13,7 +13,6 @@ import xbot.common.controls.Lidar;
 //import xbot.common.controls.*;
 import xbot.common.controls.Nav6Gyro;
 import xbot.common.controls.AnalogHIDButton.AnalogHIDDescription;
-import xbot.common.injection.MockRobotIO;
 import xbot.common.properties.PropertyManager;
 import xbot.common.wpi_extensions.mechanism_wrappers.AdvancedJoystickButton;
 import xbot.common.wpi_extensions.mechanism_wrappers.AnalogInputWPIAdapater;
@@ -37,6 +36,7 @@ import xbot.common.wpi_extensions.mechanism_wrappers.XServo;
 import xbot.common.wpi_extensions.mechanism_wrappers.XSolenoid;
 import xbot.common.wpi_extensions.mechanism_wrappers.XSpeedController;
 import edu.wpi.first.wpilibj.MockGyro;
+import edu.wpi.first.wpilibj.MockRobotIO;
 import edu.wpi.first.wpilibj.I2C.Port;
 import edu.wpi.first.wpilibj.livewindow.LiveWindow;
 import edu.wpi.first.wpilibj.livewindow.LiveWindowSendable;
