+++ tests/edu/wpi/first/wpilibj/MockDistanceSensor.java
@@ -1,7 +1,6 @@
 package edu.wpi.first.wpilibj;
 
 import xbot.common.controls.DistanceSensor;
-import xbot.common.injection.MockRobotIO;
 
 public class MockDistanceSensor implements DistanceSensor {
