+++ tests/edu/wpi/first/wpilibj/MockGyro.java
@@ -1,6 +1,5 @@
 package edu.wpi.first.wpilibj;
 
-import xbot.common.injection.MockRobotIO;
 import xbot.common.math.ContiguousDouble;
 import xbot.common.wpi_extensions.mechanism_wrappers.XGyro;
