+++ tests/edu/wpi/first/wpilibj/MockSpeedController.java
@@ -2,7 +2,6 @@ package edu.wpi.first.wpilibj;
 
 import org.apache.log4j.Logger;
 
-import xbot.common.injection.MockRobotIO;
 import xbot.common.wpi_extensions.mechanism_wrappers.XSpeedController;
 
 public class MockSpeedController implements XSpeedController {
