+++ tests/xbot/common/injection/BaseWPITest.java
@@ -10,6 +10,7 @@ import com.google.inject.Injector;
 import xbot.common.properties.PropertyManager;
 import edu.wpi.first.wpilibj.HLUsageReporting;
 import edu.wpi.first.wpilibj.MockHLUsageReporting;
+import edu.wpi.first.wpilibj.MockRobotIO;
 import edu.wpi.first.wpilibj.MockRobotState;
 import edu.wpi.first.wpilibj.MockTimer;
 import edu.wpi.first.wpilibj.RobotState;
