+++ tests/xbot/common/injection/wpi_factories/MockWPIFactory.java
@@ -7,7 +7,6 @@ import com.google.inject.Inject;
 import xbot.common.controls.AnalogHIDButton;
 import xbot.common.controls.AnalogHIDButton.AnalogHIDDescription;
 import xbot.common.controls.DistanceSensor;
-import xbot.common.injection.MockRobotIO;
 import xbot.common.injection.wpi_factories.WPIFactory;
 import xbot.common.wpi_extensions.mechanism_wrappers.AdvancedJoystickButton;
 import xbot.common.wpi_extensions.mechanism_wrappers.XAnalogInput;
@@ -31,6 +30,7 @@ import edu.wpi.first.wpilibj.MockEncoder;
 import edu.wpi.first.wpilibj.MockGyro;
 import edu.wpi.first.wpilibj.MockJoystick;
 import edu.wpi.first.wpilibj.MockPowerDistributionPanel;
+import edu.wpi.first.wpilibj.MockRobotIO;
 import edu.wpi.first.wpilibj.MockServo;
 import edu.wpi.first.wpilibj.MockSolenoid;
 import edu.wpi.first.wpilibj.MockSpeedController;
